# Improve encoding performance.